bugfix(server,frontend): fix SPA routing and refactor layout into modular components (#607)
## Summary
SPA client-side routing was broken for nested paths (e.g.,
`/provider/settings/detail/123`) because a greedy `/:page` route
intercepted them before `NoRoute` could serve `index.html`.

The layout file had also grown unwieldy at ~1000 lines, making it hard
to maintain.

### Major
- Removed the greedy `/:page` catch-all route; SPA fallback now relies
solely on `NoRoute`, ensuring all unmatched frontend paths correctly
serve `index.html` for React Router
- `UseIndexHTML` now reads from the embedded filesystem instead of using
`c.FileFromFS`, and sets proper no-cache headers to prevent stale HTML
caching
- Extracted `Layout.tsx` (~1000 lines) into focused modules:
`ActivityBar.tsx`, `Sidebar.tsx`, `useActivityItems.tsx`,
`constants.ts`, `types.ts`
- Added `icon.png` and cleaned up `icon.svg` (removed Inkscape metadata)

### Minor
- Static files (`icon.svg`, `icon.png`) now served via explicit `GET`
handlers with proper content types instead of `StaticFile`
- Moved SPA catch-all comment to clarify registration order intent
