+++ frontend/public/assets/mockServiceWorker.js
